+++ Environment_Engine/Compute/MergePanels.cs
