+++ Environment_Engine/Environment_Engine.csproj
@@ -78,24 +78,39 @@
   <ItemGroup>
     <Compile Include="Compute\BooleanIntersect.cs" />
     <Compile Include="Compute\EquipmentGain.cs" />
+    <Compile Include="Compute\JulianDay.cs" />
     <Compile Include="Compute\LightingGain.cs" />
     <Compile Include="Compute\Occupancy.cs" />
     <Compile Include="Compute\PeopleGain.cs" />
     <Compile Include="Compute\PlugGain.cs" />
     <Compile Include="Compute\SolarPosition.cs" />
+    <Compile Include="Compute\ThirdOrderPolynomial.cs" />
     <Compile Include="Convert\Degrees.cs" />
     <Compile Include="Convert\Edge.cs" />
     <Compile Include="Create\LightReflectanceFragment.cs" />
     <Compile Include="Create\LightTransmittanceFragment.cs" />
     <Compile Include="Create\Location.cs" />
     <Compile Include="Create\SpaceTime.cs" />
+    <Compile Include="Modify\AddOpenings.cs" />
+    <Compile Include="Modify\ChangeAdjacentSpace.cs" />
+    <Compile Include="Modify\ChangeAdjacentSpaces.cs" />
     <Compile Include="Modify\CleanPanel.cs" />
     <Compile Include="Modify\FixNormal.cs" />
     <Compile Include="Modify\OffsetOpenings.cs" />
-    <Compile Include="Modify\Remove.cs" />
-    <Compile Include="Modify\RemoveOpenings.cs" />
+    <Compile Include="Modify\RemoveOpening.cs" />
+    <Compile Include="Modify\RemovePanel.cs" />
+    <Compile Include="Modify\RemoveOpeningsByName.cs" />
+    <Compile Include="Modify\RemovePanels.cs" />
+    <Compile Include="Modify\SetAdjacentSpaces.cs" />
+    <Compile Include="Modify\SetFloorPanels.cs" />
+    <Compile Include="Modify\SetOpeningConstruction.cs" />
     <Compile Include="Modify\SetOpeningType.cs" />
+    <Compile Include="Modify\SetPanelTypeByAdjacencies.cs" />
+    <Compile Include="Modify\SetRoofPanels.cs" />
+    <Compile Include="Modify\SetShadePanels.cs" />
+    <Compile Include="Modify\SetWallPanels.cs" />
     <Compile Include="Modify\SplitOpening.cs" />
+    <Compile Include="Modify\UpdatePanelTypeByCustomData.cs" />
     <Compile Include="Query\Absorptance.cs" />
     <Compile Include="Query\IsValidLocation.cs" />
     <Compile Include="Query\DateTime.cs" />
@@ -130,11 +145,11 @@
     <Compile Include="Create\IElement2D\NewInternalElement2D.cs" />
     <Compile Include="Create\SimulationResult.cs" />
     <Compile Include="Create\SpaceAnalyticalFragment.cs" />
-    <Compile Include="Modify\AdjacentSpace.cs" />
+    <Compile Include="Modify\AddAdjacentSpace.cs" />
     <Compile Include="Modify\AddOpening.cs" />
-    <Compile Include="Compute\MergeBuildingElements.cs" />
+    <Compile Include="Compute\MergePanels.cs" />
     <Compile Include="Modify\Copy.cs" />
-    <Compile Include="Modify\SetConstruction.cs" />
+    <Compile Include="Modify\SetConstructions.cs" />
     <Compile Include="Modify\SetPanelType.cs" />
     <Compile Include="Modify\AddFragment.cs" />
     <Compile Include="Modify\SetInternalElements2D.cs" />
