+++ Environment_Engine/Modify/AddAdjacentSpace.cs
@@ -64,57 +64,6 @@ namespace BH.Engine.Environment
             Panel clonedPanel = panel.DeepClone<Panel>();
             return AddAdjacentSpace(new List<Panel> { clonedPanel }, spaceName)[0];
         }
-
-        [Description("Returns a single Environment Panel with the provided space names added as the connecting spaces")]
-        [Input("panel", "A single Environment Panel to add the space names to")]
-        [Input("spaceNames", "The collection of names of the spaces the panel is connected to")]
-        [Output("panel", "A modified Environment Panel with the provided space names listed as the connecting spaces")]
-        public static Panel SetAdjancentSpaces(this Panel panel, List<string> spaceNames)
-        {
-            Panel clonedPanel = panel.DeepClone<Panel>();
-            clonedPanel.ConnectedSpaces = spaceNames;
-            return clonedPanel;
-        }
-
-        [Description("Returns a single Environment Panel with an updated connected space name")]
-        [Input("panel", "A single Environment Panel to change the connected space name of")]
-        [Input("spaceNameToChange", "The space name to replace")]
-        [Input("replacementSpaceName", "The new space name to use")]
-        [Output("panel", "A modified Environment Panel with the changed connected space name")]
-        public static Panel ChangeAdjacentSpace(this Panel panel, string spaceNameToChange, string replacementSpaceName)
-        {
-            Panel clonedPanel = panel.DeepClone<Panel>();
-            for(int x = 0; x < clonedPanel.ConnectedSpaces.Count; x++)
-            {
-                if (clonedPanel.ConnectedSpaces[x] == spaceNameToChange)
-                    clonedPanel.ConnectedSpaces[x] = replacementSpaceName;
-            }
-
-            return clonedPanel;
-        }
-
-        [Description("Returns a collection of Environment Panels where any connected spaces which are detailed within the spaceNamesToChange are replaced by a replacementSpaceName. The spaceNamesToChange and replacementSpaceNames should match length to provide a 1:1 change")]
-        [Input("panels", "A collection of Environment Panels to update the connected space names of")]
-        [Input("spaceNamesToChange", "A collection of space names which should be updated")]
-        [Input("replacementSpaceNames", "A collection of space names to replace with")]
-        [Output("panels", "A collection of Environment Panels modified so that space names are changed as appropriate")]
-        public static List<Panel> ChangeAdjacentSpaces(this List<Panel> panels, List<string> spaceNamesToChange, List<string> replacementSpaceNames)
-        {
-            List<Panel> clonedPanels = new List<Panel>(panels.Select(x => x.DeepClone<Panel>()).ToList());
-            if(spaceNamesToChange.Count != replacementSpaceNames.Count)
-            {
-                BH.Engine.Reflection.Compute.RecordError("Please ensure the number of replacement space names matches the number of changing space names. Panels returned without change");
-                return clonedPanels;
-            }
-
-            for(int x = 0; x < spaceNamesToChange.Count; x++)
-            {
-                for (int a = 0; a < clonedPanels.Count; a++)
-                    clonedPanels[a] = ChangeAdjacentSpace(clonedPanels[a], spaceNamesToChange[x], replacementSpaceNames[x]);
-            }
-
-            return clonedPanels;
-        }
     }
 }
