+++ Environment_Engine/Modify/AddOpenings.cs
@@ -1,0 +1,76 @@
+﻿/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using System.Linq;
+using System.Collections.Generic;
+using BH.oM.Environment.Elements;
+using BH.oM.Geometry;
+using BH.Engine.Geometry;
+
+using BH.oM.Reflection.Attributes;
+using System.ComponentModel;
+using BH.Engine.Base;
+
+namespace BH.Engine.Environment
+{
+    public static partial class Modify
+    {
+        /***************************************************/
+        /**** Public Methods                            ****/
+        /***************************************************/
+
+        [Deprecated("2.4", "Deprecated to expose tolerance as optional parameter for greater control", null, "AddOpenings(this List<Panel> panels, List<Opening> openings, double tolerance = Tolerance.Distance)")]
+        public static List<Panel> AddOpenings(this List<Panel> panels, List<Opening> openings)
+        {
+            return panels.AddOpenings(openings, BH.oM.Geometry.Tolerance.Distance);
+        }
+
+        [Description("Returns a list of Environment Panel with the provided openings added. Openings are added to the panels which contain them geometrically.")]
+        [Input("panels", "A collection of Environment Panels to add the opening to")]
+        [Input("openings", "A collection of Environment Openings to add to the panels")]
+        [Input("centroidTolerance", "Set the tolerance for obtaining the centroid of openings, default is set to BH.oM.Geometry.Tolerance.Distance")]
+        [Input("containingTolerance", "Set the tolerance for determining geometric association of openings to panels, default is set to BH.oM.Geometry.Tolerance.Distance")]
+        [Output("panels", "A collection of modified Environment Panels with the provided openings added")]
+        public static List<Panel> AddOpenings(this List<Panel> panels, List<Opening> openings, double centroidTolerance = BH.oM.Geometry.Tolerance.Distance, double containingTolerance = BH.oM.Geometry.Tolerance.Distance)
+        {
+            List<Panel> clonedPanels = new List<Panel>(panels.Select(x => x.DeepClone<Panel>()).ToList());
+            List<Opening> clonedOpenings = new List<Opening>(openings.Select(x => x.DeepClone<Opening>()).ToList());
+
+            foreach (Opening o in clonedOpenings)
+            {
+                Point centre = o.Polyline().Centroid(centroidTolerance);
+                if (centre != null)
+                {
+                    Panel panel = clonedPanels.PanelsContainingPoint(centre, false, containingTolerance).FirstOrDefault();
+                    if (panel != null)
+                    {
+                        if (panel.Openings == null) panel.Openings = new List<Opening>();
+                        panel.Openings.Add(o);
+                    }
+                }
+            }
+
+            return clonedPanels;
+        }
+    }
+}
+
