+++ Environment_Engine/Modify/ChangeAdjacentSpaces.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -20,14 +20,13 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using System.Linq;
 using System.Collections.Generic;
 using BH.oM.Environment.Elements;
-using BH.oM.Geometry;
-using BH.Engine.Geometry;
 
 using BH.oM.Reflection.Attributes;
 using System.ComponentModel;
+
+using System.Linq;
 using BH.Engine.Base;
 
 namespace BH.Engine.Environment
@@ -38,16 +37,27 @@ namespace BH.Engine.Environment
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Returns a single Environment Panel with the provided opening. Opening is added to the provided panel regardless of geometric association")]
-        [Input("panel", "A single Environment Panel to add the opening to")]
-        [Input("opening", "The Environment Opening to add to the panel")]
-        [Output("panel", "A modified Environment Panel with the provided opening added")]
-        public static Panel AddOpening(this Panel panel, Opening opening)
+        [Description("Returns a collection of Environment Panels where any connected spaces which are detailed within the spaceNamesToChange are replaced by a replacementSpaceName. The spaceNamesToChange and replacementSpaceNames should match length to provide a 1:1 change")]
+        [Input("panels", "A collection of Environment Panels to update the connected space names of")]
+        [Input("spaceNamesToChange", "A collection of space names which should be updated")]
+        [Input("replacementSpaceNames", "A collection of space names to replace with")]
+        [Output("panels", "A collection of Environment Panels modified so that space names are changed as appropriate")]
+        public static List<Panel> ChangeAdjacentSpaces(this List<Panel> panels, List<string> spaceNamesToChange, List<string> replacementSpaceNames)
         {
-            Panel clone = panel.DeepClone<Panel>();
-            if (clone.Openings == null) clone.Openings = new List<Opening>();
-            clone.Openings.Add(opening);
-            return clone;
+            List<Panel> clonedPanels = new List<Panel>(panels.Select(x => x.DeepClone<Panel>()).ToList());
+            if (spaceNamesToChange.Count != replacementSpaceNames.Count)
+            {
+                BH.Engine.Reflection.Compute.RecordError("Please ensure the number of replacement space names matches the number of changing space names. Panels returned without change");
+                return clonedPanels;
+            }
+
+            for (int x = 0; x < spaceNamesToChange.Count; x++)
+            {
+                for (int a = 0; a < clonedPanels.Count; a++)
+                    clonedPanels[a] = ChangeAdjacentSpace(clonedPanels[a], spaceNamesToChange[x], replacementSpaceNames[x]);
+            }
+
+            return clonedPanels;
         }
     }
 }
