+++ Environment_Engine/Modify/SetPanelType.cs
@@ -40,79 +40,6 @@ namespace BH.Engine.Environment
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Calculates the Panel type by the spaces adjacent to it. This is only valid for wall panels and is NOT valid for roof or floor panels")]
-        [Input("panels", "A collection of Environment Panels to calculate the type of")]
-        [Output("panels", "A collection of Environment Panels with their type set")]
-        public static List<Panel> SetPanelTypeByAdjacencies(this List<Panel> panels)
-        {
-            List<Panel> clones = new List<Panel>(panels.Select(x => x.DeepClone<Panel>()).ToList());
-
-            foreach(Panel panel in clones)
-            {
-                if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 0)
-                    panel.Type = PanelType.Shade;
-                else if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 1)
-                    panel.Type = PanelType.WallExternal;
-                else if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 2)
-                    panel.Type = PanelType.WallInternal;                
-            }
-
-            return clones;
-        }
-
-        [Description("Sets the panel type based on some custom data for the panel")]
-        [Input("panels", "A collection of Environment Panels to set the type of")]
-        [Input("customDataKey", "The key of the custom data dictionary the algorithm should use to find the custom data information to set the panel type from, default empty string")]
-        [Output("panels", "A collection of Environment Panels with their type set")]
-        public static List<Panel> UpdatePanelTypeByCustomData(this List<Panel> panels, string customDataKey = "")
-        {
-            List<Panel> clones = new List<Panel>(panels.Select(x => x.DeepClone<Panel>()).ToList());
-            if (customDataKey == "")
-            {
-                BH.Engine.Reflection.Compute.RecordWarning("You did not set a custom data key and so no modifications were made to these Panels");
-                return clones; //Make no changes if no one has presented a key
-            }
-
-            foreach(Panel panel in clones)
-            {
-                if(panel.CustomData.ContainsKey(customDataKey))
-                {
-                    PanelType pType = PanelType.Undefined;
-                    string type = panel.CustomData[customDataKey] as string;
-                    type = type.ToLower();
-
-                    if (type == "underground wall")
-                        pType = PanelType.UndergroundWall;
-                    else if (type == "curtain wall")
-                        pType = PanelType.CurtainWall;
-                    else if (type == "external wall")
-                        pType = PanelType.WallExternal;
-                    else if (type == "internal wall")
-                        pType = PanelType.WallInternal;
-                    else if (type == "no type")
-                        pType = PanelType.Undefined;
-                    else if (type == "shade")
-                        pType = PanelType.Shade;
-                    else if (type == "solar/pv panel")
-                        pType = PanelType.SolarPanel;
-                    else if (type == "roof")
-                        pType = PanelType.Roof;
-                    else if (type == "underground ceiling")
-                        pType = PanelType.UndergroundCeiling;
-                    else if (type == "internal floor")
-                        pType = PanelType.FloorInternal;
-                    else if (type == "exposed floor")
-                        pType = PanelType.FloorExposed;
-                    else if (type == "slab on grade")
-                        pType = PanelType.SlabOnGrade;
-
-                    panel.Type = pType;
-                }
-            }
-
-            return clones;
-        }
-
         [Description("Sets the panel type based on the provided type")]
         [Input("panels", "A collection of Environment Panels to set the type of")]
         [Input("type", "The panel type to assign to the panels")]
@@ -124,129 +51,6 @@ namespace BH.Engine.Environment
                 p.Type = type;
 
             return clones;
-        }
-
-        [Description("Returns the floor panels of a space represented by Environment Panels")]
-        [Input("panelsAsSpace", "A collection of Environment Panels that represent a closed space")]
-        [Output("floorPanels", "BHoM Environment panel representing the floor of the space")]
-        public static List<Panel> SetFloorPanels(this List<Panel> panelsAsSpace)
-        {
-            List<Panel> clones = new List<Panel>(panelsAsSpace.Select(x => x.DeepClone<Panel>()).ToList());
-
-            //Find the panel(s) that are at the lowest point of the space...
-            double minZ = 1e10;
-            foreach (Panel panel in clones)
-            {
-                if (panel.MinimumLevel() == panel.MaximumLevel())
-                    minZ = Math.Min(minZ, panel.MinimumLevel());
-            }
-
-            List<Panel> floorPanels = clones.Where(x => x.MinimumLevel() == minZ && x.MaximumLevel() == minZ).ToList();
-
-            if (floorPanels.Count == 0)
-            {
-                BH.Engine.Reflection.Compute.RecordWarning("Could not find floor panel");
-                return null;
-            }
-
-            foreach (Panel panel in floorPanels)
-            {
-                if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 1)
-                    panel.Type = PanelType.SlabOnGrade;
-                else if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 2)
-                    panel.Type = PanelType.FloorInternal;
-            }
-
-
-            return floorPanels;
-        }
-
-        [Description("Returns the roof panels of a space represented by Environment Panels")]
-        [Input("panelsAsSpace", "A collection of Environment Panels that represent a closed space")]
-        [Output("roofPanels", "BHoM Environment panel representing the roof of the space")]
-        public static List<Panel> SetRoofPanels(this List<Panel> panelsAsSpace)
-        {
-            List<Panel> clones = new List<Panel>(panelsAsSpace.Select(x => x.DeepClone<Panel>()).ToList());
-
-            //Find the panel(s) that are at the highest point of the space...
-            double minZ = 1e10;
-            double maxZ = -1e10;
-            foreach (Panel panel in clones)
-            {
-                if (panel.MinimumLevel() == panel.MaximumLevel())
-                {
-                    minZ = Math.Min(minZ, panel.MinimumLevel());
-                    maxZ = Math.Max(maxZ, panel.MaximumLevel());
-                }
-            }
-
-            List<Panel> roofPanels = clones.Where(x => ((x.MaximumLevel() != minZ) && (Math.Round(x.Tilt()) >= 92 || Math.Round(x.Tilt()) <= 88))).ToList();
-
-            foreach (Panel panel in roofPanels)
-            {
-                if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 1)
-                    panel.Type = PanelType.Roof;
-                else if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 2)
-                    panel.Type = PanelType.Ceiling;
-            }
-
-            foreach (Panel panel in roofPanels)
-            {
-                if (panel.Type == PanelType.Ceiling && panel.MaximumLevel() != maxZ)
-                    panel.Type = PanelType.FloorInternal;
-            }
-
-            return roofPanels;
-        }
-
-        [Description("Returns the wall panels of a space represented by Environment Panels and fixes PanelType")]
-        [Input("panelsAsSpace", "A collection of Environment Panels that represent a closed space")]
-        [Output("wallPanels", "BHoM Environment panel representing the wall of the space")]
-        public static List<Panel> SetWallPanels(this List<Panel> panelsAsSpace)
-        {
-            List<Panel> clones = new List<Panel>(panelsAsSpace.Select(x => x.DeepClone<Panel>()).ToList());
-
-            //Find the panel(s) that are horizontal
-            double minZ = 1e10;
-            foreach (Panel panel in clones)
-            {
-                if (panel.MinimumLevel() == panel.MaximumLevel())
-                    minZ = Math.Min(minZ, panel.MinimumLevel());
-            }
-
-            List<Panel> wallPanels = clones.Where(x => x.Tilt() < 92 && x.Tilt() > 88).ToList();
-
-            if (wallPanels.Count == 0)
-            {
-                BH.Engine.Reflection.Compute.RecordWarning("Could not find wall panel");
-                return null;
-            }
-
-            foreach (Panel panel in wallPanels)
-            {
-                if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 1)
-                    panel.Type = PanelType.WallExternal;
-                else if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 2)
-                    panel.Type = PanelType.WallInternal;
-            }
-
-            return wallPanels;
-        }
-
-        [Description("Returns the shade panels represented by Environment Panels with no adj spaces and fixes PanelType")]
-        [Input("panels", "A collection of Environment Panels")]
-        [Output("shadePanels", "BHoM Environment panel representing the shade")]
-        public static List<Panel> SetShadePanels(this List<Panel> panels)
-        {
-            //Find the panel(s) without connected spaces and set as shade
-            List<Panel> shadePanels = new List<Panel>(panels.Select(x => x.DeepClone<Panel>()).ToList()); 
-            foreach (Panel panel in shadePanels)
-            {
-                if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 0)
-                    panel.Type = PanelType.Shade;
-            }
-
-            return shadePanels;
         }
     }
 }
