+++ Environment_Engine/Modify/SetPanelTypeByAdjacencies.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -25,9 +25,11 @@ using System.Collections.Generic;
 using BH.oM.Environment.Elements;
 using BH.oM.Geometry;
 using BH.Engine.Geometry;
-
+using BH.oM.Environment.Fragments;
+using System;
 using BH.oM.Reflection.Attributes;
 using System.ComponentModel;
+
 using BH.Engine.Base;
 
 namespace BH.Engine.Environment
@@ -38,16 +40,24 @@ namespace BH.Engine.Environment
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Returns a single Environment Panel with the provided opening. Opening is added to the provided panel regardless of geometric association")]
-        [Input("panel", "A single Environment Panel to add the opening to")]
-        [Input("opening", "The Environment Opening to add to the panel")]
-        [Output("panel", "A modified Environment Panel with the provided opening added")]
-        public static Panel AddOpening(this Panel panel, Opening opening)
+        [Description("Calculates the Panel type by the spaces adjacent to it. This is only valid for wall panels and is NOT valid for roof or floor panels")]
+        [Input("panels", "A collection of Environment Panels to calculate the type of")]
+        [Output("panels", "A collection of Environment Panels with their type set")]
+        public static List<Panel> SetPanelTypeByAdjacencies(this List<Panel> panels)
         {
-            Panel clone = panel.DeepClone<Panel>();
-            if (clone.Openings == null) clone.Openings = new List<Opening>();
-            clone.Openings.Add(opening);
-            return clone;
+            List<Panel> clones = new List<Panel>(panels.Select(x => x.DeepClone<Panel>()).ToList());
+
+            foreach (Panel panel in clones)
+            {
+                if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 0)
+                    panel.Type = PanelType.Shade;
+                else if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 1)
+                    panel.Type = PanelType.WallExternal;
+                else if (panel.ConnectedSpaces.Where(x => x != "-1").ToList().Count == 2)
+                    panel.Type = PanelType.WallInternal;
+            }
+
+            return clones;
         }
     }
 }
